+++ client/CHANGELOG.md
@@ -15,6 +15,7 @@ All notable changes to this project will be documented in this file.  The format
 
 ### Added
 * Add support for retrieving historical auction information via the addition of an optional `--block-identifier` arg in the `get-auction-info` subcommand.
+* Add `keygen::generate_files` to FFI.
 
 ### Changed
 * Change `account-address` subcommand to output properly formatted string.
@@ -22,7 +23,6 @@ All notable changes to this project will be documented in this file.  The format
 * Change `make-deploy`, `make-transfer` and `sign-deploy` to not overwrite files unless `--force` is passed.
 * Change `make-deploy`, `make-transfer` and `sign-deploy` to use transactional file writing for enhanced safety and reliability.
 * Update pinned version of Rust to `nightly-2021-06-17`
-
 
 
 ## [1.2.0] - 2021-05-27